[occm] multizone race condition
We cannot definitively determine whether a node exists within our zone without the provider ID string.
